Be more diligent about freeing up resources when downloading responses
If we download a large response, the ByteArrayOutputStream will
remain as large as the response until that client is disconnected,
and garbage collected. Rather than reusing BAOS, create new ones
and allow big byte[] arrays to be recycled.

Also allow modification of response headers in the
"responseHeaderReceived" method. Of course, you can shoot yourself
in the foot doing this. Don't do anything to affect how the response
body s read!