Allow the response content recorded to be size limited.
This allows the caller to prevent recording of excessively sized responses, for example, streaming video. If a response content is larger than the specified size, an IOException will be thrown *after* the entire response has been read.

Note that response contents that are being simultaneously streamed to an OutputStream using HttpClient.fetchResponseContent(OutputStream) wil be completely written to that OutputStream before the IOException is thrown.