Facilitate loop avoidance.
If a user registers all related servers in a ServerGroup instance,
then uses a LoopAvoidingTargetedConnectionHandler as a wrapper around
the desired TargetedConnectionHandler, it will reject any connections
to one of the registered Servers.

Note that it cannot avoid loops if the target is specified within the
protocol i.e. if it is null when passed to the TargetedConnectionHandler.

In such a case, implementations should check with the ServerGroup once
the target has been established in order to avoid loops.