Provide useful methods for buffering StreamingMessages and streaming BufferedMessages
Also implement methods for copying StreamingMessages as they are streamed. This is
useful for performance reasons, in that the message content can be streamed to the
recipient as it is read from the client/server, rather than having to buffer it
completely on the proxy before being able to stream it to the reciient.

This also enables setting limits on the maximum size of message bodies which will be
recorded.