Make use of the new delayedCopy methods, and specify a maximum message content size
Messages that overflow the content size will be stored with no content at all.
It would be easy to allow the user to choose whether to record the allowed
max size of content rather than none at all, but interpreting that when the content
is read back would be tricky. Perhaps the DAO should be extended to allow recording
of messages where the content overflowed. Getting that back from the DAO may prove tricky,
unfortunately. I guess it would have to be a property of the Conversation . . .