Add an implementation of a BufferingHttpRequestHandler
This implementation allows the user to choose the maximum size of message content to buffer,
and gracefully handles cases where the message content is larger than the specified maximum.

Users should override the processRequest() and processResponse() methods in order to gain access
to the BufferedMessage's. Implementations can also choose whether to buffer specific responses, based
on the request info, as well as the response headers.

Finally, implementations will be notified if the message body ends up being larger than the specified
maximum, and will be provided with the request and response headers respectively.