Implement initial untested support for "100 Continue" responses
If the request includes an "Expect: continue" header, we only submit the request headers
to the server, and read the response. If the response is "100 Continue",
create a new request with the original request header, and the request content, and reissue
the request.

DefaultHttpClient has been modified to only invoke Client.sendRequestContent() is "isContinue"
is true, and Client has been modified maintain its state appropriately.