Implement an AJP (Apache Jserv Protocol) 1.3 compatible client
This allows for direct requests to be made to e.g. a Tomcat server on port 8009
(typically), potentially bypassing any verification/validation performed
by the web server fronting the tomcat instance.

For example, one can specify an X509Certificate to present to the Tomcat instance,
even if one does not possess the corresponding private key.

One can also specify attributes to be set on the HttpServletRequest object, which,
while they may not show up in request.getAttributeNames(), are present if you do
request.getAttribute(name). Note that you can only specify String values, not other
object types.

Finally, one can also specify the remote address (i.e. the source of the request),
enabling a remote attacker to access pages that are restricted to localhost, for instance.