Alter the BufferedMessageInterceptor to pass a BufferedRequest to the various response methods
If you want to process the response, you really want to have the full request, not just the header.

NOTE: This change also changes the behaviour relating to IGNOREd requests. If the request is IGNORED,
then no further methods on this class will be called for that conversation, since the BufferedRequest
is not available. This is conceptually cleaner too, since if you ignore the request, surely the response
is also not interesting.